moved getUserData to firebase/users.ts
