+++ components/pages/LocationPage.tsx
@@ -1,11 +1,12 @@
 import { useNavigation } from "@react-navigation/native";
 import { getAuth } from "firebase/auth";
-import { doc, getDoc, setDoc } from "firebase/firestore";
+import { doc, setDoc } from "firebase/firestore";
 import { useState } from "react";
 import { Text, ActivityIndicator } from "react-native";
 import { SafeAreaView } from "react-native-safe-area-context";
 import { StreamChat } from "stream-chat";
 
+import { getUserData } from "../../firebase/users";
 import { db, getStreamUserToken } from "../../firebaseConfig";
 import Colors from "../../settings/Colors";
 import TERMS from "../../settings/Terms";
@@ -22,22 +23,6 @@ const LocationPage = () => {
   const [loading, setLoading] = useState(false);
 
   const disabled = location.length === 0;
-
-  const getUserData = async () => {
-    const auth = getAuth();
-    const userId = auth.currentUser?.uid;
-    if (!userId) throw new Error("No user ID found!");
-
-    const userRef = doc(db, "users", userId);
-    const userSnap = await getDoc(userRef);
-
-    if (userSnap.exists()) {
-      return userSnap.data();
-    } else {
-      console.log("No data for user!");
-      return null;
-    }
-  };
 
   const handleSubmitLocation = async () => {
     setLoading(true);
