+++ firebase/users.ts
@@ -45,4 +45,20 @@ const addUserDetails = async (userDetails) => {
   await setDoc(userRef, userDetails, { merge: true });
 };
 
-export { fetchUsersCities, addUserCity, addUserDetails };
+const getUserData = async () => {
+  const auth = getAuth();
+  const userId = auth.currentUser?.uid;
+  if (!userId) throw new Error("No user ID found!");
+
+  const userRef = doc(db, "users", userId);
+  const userSnap = await getDoc(userRef);
+
+  if (userSnap.exists()) {
+    return userSnap.data();
+  } else {
+    console.log("No data for user!");
+    return null;
+  }
+};
+
+export { fetchUsersCities, addUserCity, addUserDetails, getUserData };
